Remove EDH teams api methods & tests
